generate documentation from config.sample.php
