+++ composer.json
@@ -17,9 +17,8 @@
         }
     },
     "require": {
-        "php": ">=5.6"
-    },
-    "require": {
-        "phpdocumentor/reflection-docblock": "~2.0"
+        "php": ">=5.6",
+        "phpdocumentor/reflection-docblock": "~2.0",
+        "symfony/console": "^3.2"
     }
 }
